No 2

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,4 +1,5 @@
 exec { 'pkill':
   command  => 'pkill killmenow',
   provider => 'shell',
+  onlyif   => 'pgrep killmenow',
 }
